+++ recipes/default.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: dropbox
 # Recipe:: default
 #
-# Copyright 2014 Jonathan Hartman
+# Copyright 2014-2015 Jonathan Hartman
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,5 +19,5 @@
 #
 
 dropbox 'dropbox' do
-  package_url node['dropbox']['package_url']
+  source node['dropbox']['source']
 end