+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Dropbox Cookbook CHANGELOG
 
 v?.?.? (????-??-??)
 -------------------
+- Rewrite to modern resource/provider standards (breaking compatibility with
+  Chef < 12)
+- Change `package_url` attribute to `source`
 
 v0.1.0 (2014-12-27)
 -------------------